refactor(tool): establish contract package
- move runtime tool contracts under core.tool
- move shared tool execution under core.tool
- migrate runtime and test imports to canonical paths
