+++ core/__init__.py
@@ -41,16 +41,16 @@ _EXPORT_MODULES = {
     "TurnStartEvent": "core.events",
     "runtime_event_from_tuple": "core.events",
     "tuple_payload_from_event": "core.events",
-    "BeforeToolCallResult": "core.execution",
-    "ToolExecutionHooks": "core.execution",
-    "ToolExecutionPatch": "core.execution",
-    "ToolExecutionRequest": "core.execution",
-    "ToolExecutionResult": "core.execution",
-    "execute_tool_calls": "core.execution",
-    "execute_tools": "core.execution",
-    "public_tool_input": "core.execution",
-    "summarise": "core.execution",
-    "tool_source": "core.execution",
+    "BeforeToolCallResult": "core.tool.execution",
+    "ToolExecutionHooks": "core.tool.execution",
+    "ToolExecutionPatch": "core.tool.execution",
+    "ToolExecutionRequest": "core.tool.execution",
+    "ToolExecutionResult": "core.tool.execution",
+    "execute_tool_calls": "core.tool.execution",
+    "execute_tools": "core.tool.execution",
+    "public_tool_input": "core.tool.execution",
+    "summarise": "core.tool.execution",
+    "tool_source": "core.tool.execution",
     "LLMInvokeFailure": "core.llm_invoke_errors",
     "classify_llm_invoke_failure": "core.llm_invoke_errors",
     "AppRuntimeMessage": "core.messages",
@@ -62,11 +62,11 @@ _EXPORT_MODULES = {
     "ProviderHooks": "core.provider",
     "ProviderRequest": "core.provider",
     "resolve_llm_api_key": "core.provider",
-    "AgentTool": "core.types",
-    "AgentToolContext": "core.types",
-    "AgentToolExecutor": "core.types",
-    "RuntimeTool": "core.types",
-    "ToolParallelism": "core.types",
+    "AgentTool": "core.tool.contracts",
+    "AgentToolContext": "core.tool.contracts",
+    "AgentToolExecutor": "core.tool.contracts",
+    "RuntimeTool": "core.tool.contracts",
+    "ToolParallelism": "core.tool.contracts",
 }
