+++ core/agent/agent.py
@@ -22,11 +22,11 @@ from core.agent.provider_hooks import ProviderHookDelegate
 from core.agent.react_loop import run_react_loop
 from core.agent.run_io import AgentRunInput, AgentRunResult
 from core.events import RuntimeEventCallback, TupleEventCallback
-from core.execution import ToolExecutionHooks
+from core.tool.execution import ToolExecutionHooks
 from core.llm.factory import LLMRole
 from core.messages import ProviderMessage, RuntimeMessage, RuntimeMessageLike
 from core.provider import ProviderHooks, ProviderRequest
-from core.types import RuntimeTool
+from core.tool.contracts import RuntimeTool
 
 if TYPE_CHECKING:
     from core.agent.goals import Goal
