+++ core/agent/react_loop.py
@@ -38,7 +38,7 @@ from core.events import (
     TurnEndEvent,
     TurnStartEvent,
 )
-from core.execution import (
+from core.tool.execution import (
     ToolExecutionHooks,
     ToolExecutionRequest,
     ToolExecutionResult,
@@ -48,7 +48,7 @@ from core.execution import (
 from core.llm.types import ToolCall
 from core.messages import MessageMapper
 from core.provider import ProviderRequest
-from core.types import RuntimeTool
+from core.tool.contracts import RuntimeTool
 from platform.observability.operations_log import record_operation
 from platform.observability.trace.redaction import redact_sensitive
 from platform.observability.trace.spans import (
