+++ core/agent_harness/turns/gather_discovery_budget.py
@@ -29,7 +29,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from core.execution import (
+from core.tool.execution import (
     BeforeToolCallResult,
     ToolExecutionHooks,
     ToolExecutionRequest,
