+++ gateway/core/middleware/approvals.py
@@ -22,7 +22,7 @@ from collections.abc import Mapping
 from dataclasses import dataclass, field
 from typing import Any, Protocol
 
-from core.execution import BeforeToolCallResult, ToolExecutionHooks, ToolExecutionRequest
+from core.tool.execution import BeforeToolCallResult, ToolExecutionHooks, ToolExecutionRequest
 from gateway.core.storage.security_audit import audit_security_action
 
 APPROVE_ACTION_ID = "opensre_approval_approve"
