+++ gateway/tests/slack/test_approvals.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import threading
 from typing import Any
 
-from core.execution import ToolExecutionRequest
+from core.tool.execution import ToolExecutionRequest
 from core.llm.types import ToolCall
 from gateway.core.middleware.approvals import (
     APPROVE_ACTION_ID,
