+++ gateway/tests/telegram/test_approvals.py
@@ -6,7 +6,7 @@ import threading
 from typing import Any
 from unittest.mock import MagicMock
 
-from core.execution import ToolExecutionRequest
+from core.tool.execution import ToolExecutionRequest
 from core.llm.types import ToolCall
 from gateway.core.middleware.approvals import (
     APPROVE_ACTION_ID,
