+++ surfaces/interactive_shell/runtime/turn_seams.py
@@ -17,7 +17,7 @@ from rich.console import Console
 from core.agent_harness import OutputSink, ToolCallingTurnResult
 from core.agent_harness.ports import AnswerRequest, GatheredEvidence
 from core.agent_harness.runtime import HeadlessAgent, TurnPlan
-from core.execution import ToolExecutionHooks
+from core.tool.execution import ToolExecutionHooks
 from surfaces.interactive_shell.session import Session
 from surfaces.interactive_shell.utils.telemetry import LlmRunInfo
