+++ tests/core/agent/orchestration/test_duplicate_action_guard.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 from typing import Any
 
 from core.agent_harness.turns.action_driver import with_duplicate_action_call_guard
-from core.execution import ToolExecutionResult
+from core.tool.execution import ToolExecutionResult
 from core.llm.types import ToolCall
 
 # One emitted tool call: name, arguments, and whether the tool would succeed.
