+++ tests/core/agent_harness/test_gather_discovery_budget.py
@@ -21,7 +21,7 @@ from core.agent_harness.turns.gather_discovery_budget import (
     is_mcp_list_tools,
     with_gather_discovery_budget,
 )
-from core.execution import ToolExecutionRequest, ToolExecutionResult
+from core.tool.execution import ToolExecutionRequest, ToolExecutionResult
 from core.llm.types import ToolCall
