+++ tests/core/runtime/test_loop.py
@@ -25,7 +25,7 @@ from core.messages import (
 )
 from core.provider import ProviderHooks
 from core.tool_framework.registered_tool import RegisteredTool
-from core.types import AgentTool, AgentToolContext
+from core.tool.contracts import AgentTool, AgentToolContext
 
 
 class FakeLLM:
