+++ tests/tools/test_github_ci_fix.py
@@ -11,7 +11,7 @@ from core.agent_harness.tools.tool_context import (
     ActionToolContext,
 )
 from core.tool_framework.registered_tool import RegisteredTool
-from core.types import AgentToolContext
+from core.tool.contracts import AgentToolContext
 from integrations.coding_agent import CodingResult
 from integrations.github.tools.ci_fix.context import (
     CiFixContext,
