+++ tools/investigation/stages/gather_evidence/agent.py
@@ -29,7 +29,7 @@ from core.provider import ProviderHooks, ProviderRequest
 from core.state import InvestigationState
 from core.state.evidence import EvidenceEntry
 from core.tool_framework.registered_tool import RegisteredTool
-from core.types import RuntimeTool
+from core.tool.contracts import RuntimeTool
 from platform.observability import debug_print
 from platform.observability import get_progress_tracker as get_tracker
 from platform.observability.trace.redaction import (
