+++ tools/system/work_items/tool.py
@@ -27,7 +27,7 @@ from core.domain.work_items import (
 )
 from core.tool_framework.metadata import SideEffectLevel
 from core.tool_framework.tool_decorator import tool
-from core.types import AgentToolContext
+from core.tool.contracts import AgentToolContext
 from platform.scheduling.scheduler.store import add_task as add_scheduled_task
 from platform.scheduling.scheduler.store import list_tasks, update_task
 from platform.scheduling.scheduler.types import Provider, ScheduledTask, TaskKind
